+++ rules/best-practices.js
@@ -47,9 +47,6 @@ module.exports = {
   // Disallow else after a return in an if
   'no-else-return': 2,
 
-  // Disallow use of labels for anything other than loops and switches
-  'no-empty-label': 2,
-
   // Disallow use of empty destructuring patterns
   'no-empty-pattern': 2,