+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All changes to this project are documented in this file. This project adheres to
 - Tweak the `README.md` file
 - Update `eslint` dependencies to ^2.0.0
 
+#### Best Practices
+- Replace `no-empty-label` with `no-labels`
+
 ### Fixed
 - Fix the compare links in the `CHANGLOG.md` file