+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All changes to this project are documented in this file. This project adheres to
 ## [Unreleased]
 ### Added
 #### Best Practices
+- Add `array-callback-return` to enforce returns in callbacks of array's methods
 - Add `no-empty-function` to disallow use of empty functions
 
 ### Changed