+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All changes to this project are documented in this file. This project adheres to
 - Add `array-callback-return` to enforce returns in callbacks of array's methods
 - Add `no-empty-function` to disallow use of empty functions
 - Add `no-extra-label` to disallow unnecessary labels
+- Add `no-implicit-globals` to disallow var and named functions in global scope
 
 #### Stylistic Issues
 - Add `newline-per-chained-call` to not enforce newline after each chained call