+++ CHANGELOG.md
@@ -13,6 +13,9 @@ All changes to this project are documented in this file. This project adheres to
 #### ECMAScript 6
 - Add `no-new-symbol` to disallow use of the new operator with the Symbol object
 
+#### Node.js and CommonJS
+- Add `no-restricted-imports` to not restrict usage of specified node imports
+
 #### Stylistic Issues
 - Add `newline-per-chained-call` to not enforce newline after each chained call